requirements.txt
av
accelerate
sentencepiece
torch
numpy
transformers
pandas
opencv-python
Pillow
ftfy
pydantic
regex
tqdm
google-generativeai
google-genai
python-dotenv